Move scheduler policies into control plane
