+++ examples/control_plane/quota-scheduler-hint-compaction-smoke.py
@@ -16,7 +16,8 @@ REPO_ROOT = Path(__file__).resolve().parents[2]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from loopx.policies.scheduler_hint import build_scheduler_hint  # noqa: E402
+from loopx.control_plane.scheduler.scheduler_hint import build_scheduler_hint  # noqa: E402
+from loopx.policies import scheduler_hint as legacy_scheduler_hint  # noqa: E402
 from loopx.quota import _scheduler_hint  # noqa: E402
 
 
@@ -251,6 +252,7 @@ def assert_cli_compact_and_detail_contract() -> None:
 
 
 def main() -> int:
+    assert legacy_scheduler_hint.build_scheduler_hint is build_scheduler_hint
     assert_compact_scheduler("active-work", payload(should_run=True))
     assert_compact_scheduler(
         "human-gate",
