+++ loopx/todo_projection.py
@@ -4,7 +4,7 @@ from datetime import datetime
 import re
 from typing import Any
 
-from .policies.monitor_todo import (
+from .control_plane.scheduler.monitor_todo import (
     monitor_todo_expires_at,
     monitor_todo_has_schedule,
     monitor_todo_is_actionable_open,
